+++ src/main/java/duke/Duke.java
@@ -1,14 +1,8 @@
 
-import java.io.FileNotFoundException;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.time.format.DateTimeParseException;
+package duke;
 import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
 import java.util.Scanner;
-import java.io.File;
-import java.time.LocalDate;
+
