+++ src/main/java/duke/Ui.java
@@ -1,4 +1,4 @@
-import java.util.ArrayList;
+package duke;
 import java.util.List;
 
 public class Ui {
